refactor folder strucure
