+++ examples/fips_oop_demo.py
@@ -2,7 +2,8 @@ from pathlib import Path
 from datetime import datetime
 import logging
 import click
-from sec_certs.dataset import FIPSDataset, FIPSAlgorithmDataset
+from sec_certs.dataset.fips import FIPSDataset
+from sec_certs.dataset.fips_algorithm import FIPSAlgorithmDataset
 from sec_certs.configuration import config
 from sec_certs.helpers import analyze_matched_algs
